[chg] 将原先CObject文件夹内容，移动至CBasic中。优化逻辑结构，修复condition无法兼容aspect的问题

+++ tutorial/MyGCondition/MyParamCondition.h
@@ -31,7 +31,7 @@ public:
             CGRAPH_PARAM_READ_CODE_BLOCK(myParam)    // 如果当前算子，跟其他相关依赖算子不存在并行关系，则参数可以直接使用，不需要加锁
             cnt = myParam->iCount;
         }
-        return (cnt % getRange());
+        return (cnt % (int)getRange());
     }
 };
